added generic favorites presenter
