+++ movs/movs/source/View/MoviesViewController.swift
@@ -1,54 +1,15 @@
 //
-//  FirstViewController.swift
+//  MoviesViewController.swift
 //  movs
 //
-//  Created by Lorien on 15/04/19.
-//  Copyright © 2019 Lorien. All rights reserved.
+//  Created by Lorien Moisyn on 18/04/19.
+//  Copyright © 2019 Auspex. All rights reserved.
 //
 
-import UIKit
+import Foundation
 
-class MoviesViewController: UIViewController, UICollectionViewDelegate, UICollectionViewDataSource {
+protocol MoviesViewController {
 
-    @IBOutlet weak var collectionView: UICollectionView!
-    
-    var presenter: MoviesPresenter!
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        collectionView.delegate = self
-        collectionView.dataSource = self
-        
-        presenter = MoviesPresenter(vc: self)
-        presenter.getMovies()
-    }
-    
-    func updateData() {
-        collectionView.reloadData()
-    }
-
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return presenter.movies.count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "MovieCell", for: indexPath) as! MovieCollectionCell
-        cell.setup(with: presenter.movies[indexPath.row])
-        return cell
-    }
-
-}
-
-extension MoviesViewController: UIScrollViewDelegate {
-    
-    func scrollViewDidScroll(_ scrollView: UIScrollView) {
-        guard !presenter.isRequesting else { return }
-        let height = scrollView.frame.size.height
-        let contentYoffset = scrollView.contentOffset.y
-        let distanceFromBottom = scrollView.contentSize.height - contentYoffset
-        if distanceFromBottom < height {
-            presenter.getNewPage()
-        }
-    }
+    func updateData()
     
 }
